refactor: Extract components from TransactionModal
nothing was changed except adding React.memo on comp
