+++ src/ducks/transactions/TransactionModal/TransactionInfos.jsx
@@ -1,0 +1,20 @@
+import React from 'react'
+
+import styles from 'ducks/transactions/TransactionModal/TransactionModal.styl'
+
+const TransactionInfo = ({ label, value }) => (
+  <div className={styles.TransactionInfo}>
+    <span className={styles.TransactionInfoLabel}>{label} :</span>
+    {value}
+  </div>
+)
+
+const TransactionInfos = ({ infos }) => (
+  <div>
+    {infos.map(({ label, value }) => (
+      <TransactionInfo key={label} label={label} value={value} />
+    ))}
+  </div>
+)
+
+export default React.memo(TransactionInfos)
